+++ src/Attributes/Accessors.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
  * Lombok PHP - Write less code!
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2022 Marcin Orlowski
+ * @copyright ©2022-2026 Marcin Orlowski
  * @license   https://opensource.org/licenses/LGPL-3.0 LGPL-3.0
  * @link      https://github.com/MarcinOrlowski/lombok-php
  */
@@ -14,6 +14,8 @@ namespace Lombok\Attributes;
 
 /**
  * Holds all configured accessors for a target object
+ *
+ * @extends \ArrayObject<string, \ReflectionProperty>
  */
 class Accessors extends \ArrayObject
 {