+++ tests/Accessors/AccessorsTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
  * Lombok PHP - Write less code!
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2022 Marcin Orlowski
+ * @copyright ©2022-2026 Marcin Orlowski
  * @license   https://opensource.org/licenses/LGPL-3.0 LGPL-3.0
  * @link      https://github.com/MarcinOrlowski/lombok-php
  */
@@ -126,6 +126,18 @@ class AccessorsTest extends TestCase
     {
         $this->expectException(PublicPropertyException::class);
         new TestEntityWithPublicProperties();
+    }
+
+    public function testExplicitMixedTypehint(): void
+    {
+        $obj = new TestEntity();
+        $val = true;
+        $obj->setMixed($val);
+        $this->assertEquals($val, $obj->getMixed());
+        $obj->setMixed($val);
+        $this->assertEquals($val, $obj->getMixed());
+        $obj->setMixed($val);
+        $this->assertEquals($val, $obj->getMixed());
     }
 
 }