+++ tests/Entities/TestEntityWithReadonlyClassPropertyAttributes.php
@@ -13,15 +13,10 @@ declare(strict_types=1);
 namespace LombokTest\Entities;
 
 use Lombok\Getter;
+use Lombok\Setter;
 
-class TestEntityWithGetterNameConflict extends \Lombok\Helper
+readonly class TestEntityWithReadonlyClassPropertyAttributes extends \Lombok\Helper
 {
-    #[Getter]
-    protected string $text = '';
-
-    public function getText(): void
-    {
-        // empty
-    }
-
+    #[Setter, Getter]
+    protected int $number;
 }