+++ CHANGES.md
@@ -4,6 +4,14 @@
 
 # CHANGE LOG #
 
+* 1.2.0 (2026-01-08)
+  * Added support for "readonly" properties
+  * Added `HelperReadonly` class to support "readonly" classes
+  * Added support for PHP 8.3, 8.4 and 8.5
+  * Improved code typehints for linter joy
+  * Updated tests
+
+
 * 1.1.1 (2023-10-10)
   * Rewrote documentation